feat: API 호출영역 useGetRecipientsCount 로 분리
